Implemented button click.
Now buttons can behavior independently on each scene by registering a MouseHandler event. Not entirely sure if this is the way I wanted to be, but it works. Clicks only works for buttons because Contains method is inside Rectangle... Must change that so you can have this effect in another GUI elements like radio buttons, sliders, etc.